changed from double to float

+++ dataset.py
@@ -6,10 +6,10 @@ import sklearn.preprocessing as P
 
 def csr_to_torchsparse(x, gpu = False):
     coo = x.tocoo()    
-    values = torch.DoubleTensor(coo.data)
+    values = torch.FloatTensor(coo.data)
     indices = torch.LongTensor(np.vstack((coo.row, coo.col)))
     size = torch.Size(coo.shape)  
-    ret = torch.sparse.DoubleTensor(indices, values, size)
+    ret = torch.sparse.FloatTensor(indices, values, size)
     if gpu:
         ret = ret.cuda()
     return ret
